refactor(reservations): split confirm-reservation command flow
