+++ apps/server/src/domain/reservations/services/commands/reserve-bike.service.ts
@@ -21,6 +21,16 @@ export type { ReserveBikeFailure, ReserveBikeInput } from "./reserve-bike/reserv
  * - bước `persist` mới ghi side effect
  *
  * Nhờ đó public entrypoint vẫn gọn, nhưng logic bên trong dễ đọc và dễ test hơn.
+ *
+ * @param input Dữ liệu đầu vào để tạo reservation hold.
+ * @param input.userId ID user yêu cầu giữ xe.
+ * @param input.bikeId ID bike cần giữ.
+ * @param input.stationId ID station mà bike phải đang thuộc về.
+ * @param input.startTime Thời điểm bắt đầu reservation.
+ * @param input.reservationOption Cách charge reservation: one-time hay subscription.
+ * @param input.subscriptionId ID subscription nếu flow dùng gói.
+ * @param input.endTime Thời điểm kết thúc hold nếu caller muốn override mặc định.
+ * @param input.now Mốc hiện tại để áp dụng blackout và scheduling outbox.
  */
 export function reserveBike(
   input: ReserveBikeInput,
